Refactored api backend to use Flask
