+++ requirements.txt
@@ -1,4 +1,10 @@
+blinker==1.6.3
+click==8.1.7
+Flask==3.0.0
+itsdangerous==2.1.2
+Jinja2==3.1.2
 joblib==1.3.2
+MarkupSafe==2.1.3
 numpy==1.26.1
 pandas==2.1.2
 python-dateutil==2.8.2
@@ -8,3 +14,4 @@ scipy==1.11.3
 six==1.16.0
 threadpoolctl==3.2.0
 tzdata==2023.3
+Werkzeug==3.0.1
